Move the Shelley run commands into their own modules
We were mostly following this pattern already. This just makes it
consistent.

So the module structure exactly follows the CLI command hierarchy.
