+++ cardano-cli/src/Cardano/CLI/Shelley/Run/StakeAddress.hs
@@ -1,0 +1,71 @@
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
+module Cardano.CLI.Shelley.Run.StakeAddress
+  ( runStakeAddressCmd
+  ) where
+
+import           Cardano.Prelude
+
+import           Control.Monad.Trans.Except (ExceptT)
+import           Control.Monad.Trans.Except.Extra (firstExceptT, newExceptT)
+
+import           Cardano.Api (StakingVerificationKey (..),
+                   readStakingVerificationKey, readVerificationKeyStakePool,
+                   shelleyDeregisterStakingAddress, shelleyDelegateStake,
+                   shelleyRegisterStakingAddress, writeCertificate)
+import           Shelley.Spec.Ledger.Keys (hashKey)
+
+import           Cardano.CLI.Key (VerificationKeyFile(..))
+import           Cardano.CLI.Ops (CliError (..))
+import           Cardano.CLI.Shelley.Parsers
+
+
+
+runStakeAddressCmd :: StakeAddressCmd -> ExceptT CliError IO ()
+
+runStakeAddressCmd (StakeKeyRegistrationCert stkKeyVerKeyFp outputFp) =
+  runStakeKeyRegistrationCert stkKeyVerKeyFp outputFp
+runStakeAddressCmd (StakeKeyDelegationCert stkKeyVerKeyFp stkPoolVerKeyFp outputFp) =
+  runStakeKeyDelegationCert stkKeyVerKeyFp stkPoolVerKeyFp outputFp
+runStakeAddressCmd (StakeKeyDeRegistrationCert stkKeyVerKeyFp outputFp) =
+  runStakeKeyDeRegistrationCert stkKeyVerKeyFp outputFp
+runStakeAddressCmd cmd = liftIO $ putStrLn $ "runStakeAddressCmd: " ++ show cmd
+
+
+--
+-- Stake address command implementations
+--
+
+runStakeKeyRegistrationCert :: VerificationKeyFile -> OutputFile -> ExceptT CliError IO ()
+runStakeKeyRegistrationCert (VerificationKeyFile vkFp) (OutputFile oFp) = do
+  StakingVerificationKeyShelley stakeVkey <-
+    firstExceptT CardanoApiError . newExceptT $ readStakingVerificationKey vkFp
+  let regCert = shelleyRegisterStakingAddress (hashKey stakeVkey)
+  firstExceptT CardanoApiError . newExceptT $ writeCertificate oFp regCert
+
+
+runStakeKeyDelegationCert
+  :: VerificationKeyFile
+  -- ^ Delegator staking verification key file.
+  -> VerificationKeyFile
+  -- ^ Delegatee stake pool verification key file.
+  -> OutputFile
+  -> ExceptT CliError IO ()
+runStakeKeyDelegationCert (VerificationKeyFile stkKey) (VerificationKeyFile poolVKey) (OutputFile outFp) = do
+  StakingVerificationKeyShelley stakeVkey <-
+    firstExceptT CardanoApiError . newExceptT $ readStakingVerificationKey stkKey
+  poolStakeVkey <- firstExceptT CardanoApiError . newExceptT $ readVerificationKeyStakePool poolVKey
+  let delegCert = shelleyDelegateStake (hashKey stakeVkey) (hashKey poolStakeVkey)
+  firstExceptT CardanoApiError . newExceptT $ writeCertificate outFp delegCert
+
+
+runStakeKeyDeRegistrationCert :: VerificationKeyFile -> OutputFile -> ExceptT CliError IO ()
+runStakeKeyDeRegistrationCert (VerificationKeyFile vkFp) (OutputFile oFp) = do
+  StakingVerificationKeyShelley stakeVkey <-
+    firstExceptT CardanoApiError . newExceptT $ readStakingVerificationKey vkFp
+  let deRegCert = shelleyDeregisterStakingAddress (hashKey stakeVkey)
+  firstExceptT CardanoApiError . newExceptT $ writeCertificate oFp deRegCert
+
